Tidy up some naming
