+++ OnnxStack.UI/Models/UpscaleModelTemplate.cs
@@ -1,0 +1,4 @@
+﻿namespace OnnxStack.UI.Models
+{
+    public record UpscaleModelTemplate(string Name, int ScaleFactor, int SampleSize);
+}
