+++ OnnxStack.UI/appsettings.json
@@ -6,32 +6,6 @@
     }
   },
   "AllowedHosts": "*",
-  "StableDiffusionConfig": {
-    "ModelSets": [
-    ]
-  },
-  "ImageUpscalerConfig": {
-    "ModelSets": [
-      {
-        "Name": "RealSR BSRGAN x4",
-        "IsEnabled": true,
-        "Channels": 3,
-        "SampleSize": 512,
-        "ScaleFactor": 4,
-        "DeviceId": 0,
-        "InterOpNumThreads": 0,
-        "IntraOpNumThreads": 0,
-        "ExecutionMode": "ORT_SEQUENTIAL",
-        "ExecutionProvider": "DirectML",
-        "ModelConfigurations": [
-          {
-            "Type": "Unet",
-            "OnnxModelPath": "D:\\Repositories\\upscaler\\SwinIR\\003_realSR_BSRGAN_DFO_s64w8_SwinIR-M_x4_GAN.onnx"
-          }
-        ]
-      }
-    ]
-  },
   "OnnxStackUIConfig": {
     "DefaultDeviceId": 0,
     "DefaultInterOpNumThreads": 0,
